services/ai-orchestrator/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.19

# Task queue + monitoring
celery[redis]>=5.4.0
redis>=5.2.0
flower>=2.0.1

# Settings / validation
pydantic>=2.10.0
pydantic-settings>=2.7.0
PyYAML>=6.0.2

# HTTP / logs / obs
httpx>=0.28.0
structlog>=24.4.0
prometheus-client>=0.21.0
sentry-sdk>=2.19.0

# LlamaIndex (reserved for future FunctionAgent/AgentWorkflow wiring;
# current orchestrator uses SDKs directly via our LLMProvider abstraction)
llama-index-core>=0.12.0

# Provider SDKs (called directly by our LLMProvider impls)
openai>=1.58.0
anthropic>=0.40.0

# MCP client SDK
mcp>=1.1.0

# Database (usage tracking)
SQLAlchemy>=2.0.0
psycopg[binary]>=3.2.0
alembic>=1.14.0

# Media processing (stubs)
Pillow>=11.0.0

# Tests
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
